PLC Programming/S7-300-Fault-Detection-Python/requirements.txt
python-snap7==2.0.2
httpx==0.27.2
python-dotenv==1.0.1
